# docs/api/desktop-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Knotra client engine boundary",
    "version": "1.0.0",
    "description": "Implemented HTTP API for the Go engine and CLI. The protocol identifier remains knotra.desktop/1. See desktop-v1.md for normative idempotency, events, security and execution rules."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787/v1",
      "description": "Local engine"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {}
  ],
  "paths": {
    "/info": {
      "get": {
        "operationId": "get_info",
        "summary": "Discover protocol and durable engine identity",
        "responses": {
          "200": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/profiles": {
      "get": {
        "operationId": "get_profiles",
        "summary": "List engine profiles",
        "responses": {
          "200": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Profile"
                      }
                    }
                  },
                  "required": [
                    "items"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/resources": {
      "get": {
        "operationId": "get_resources",
        "summary": "List engine resources",
        "responses": {
          "200": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Resource"
                      }
                    }
                  },
                  "required": [
                    "items"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/packages/validate": {
      "post": {
        "operationId": "post_packages_validate",
        "summary": "Validate package, inputs and engine admission",
        "responses": {
          "200": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    },
                    "diagnostics": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Diagnostic"
                      }
                    }
                  },
                  "required": [
                    "valid",
                    "diagnostics"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "package": {
                    "$ref": "#/components/schemas/Package"
                  },
                  "profile": {
                    "type": "string"
                  },
                  "inputs": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "artifacts": {
                    "$ref": "#/components/schemas/ArtifactBindings"
                  }
                },
                "required": [
                  "package",
                  "profile",
                  "inputs",
                  "artifacts"
                ]
              }
            }
          }
        }
      }
    },
    "/definitions": {
      "get": {
        "operationId": "get_definitions",
        "summary": "List immutable definition versions",
        "responses": {
          "200": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Definition"
                      }
                    },
                    "nextCursor": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "items",
                    "nextCursor"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "post_definitions",
        "summary": "Commit immutable package version",
        "responses": {
          "201": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "definition": {
                      "$ref": "#/components/schemas/Definition"
                    }
                  },
                  "required": [
                    "definition"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "package": {
                    "$ref": "#/components/schemas/Package"
                  }
                },
                "required": [
                  "package"
                ]
              }
            }
          }
        }
      }
    },
    "/definitions/{definitionId}": {
      "get": {
        "operationId": "get_definitions_definitionId",
        "summary": "Read exact definition snapshot",
        "responses": {
          "200": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "definition": {
                      "$ref": "#/components/schemas/Definition"
                    }
                  },
                  "required": [
                    "definition"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "definitionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            }
          }
        ]
      }
    },
    "/runs": {
      "get": {
        "operationId": "get_runs",
        "summary": "List authoritative run projections",
        "responses": {
          "200": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "post_runs",
        "summary": "Admit and accept a run",
        "responses": {
          "201": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "run": {
                      "$ref": "#/components/schemas/Run"
                    }
                  },
                  "required": [
                    "run"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "definitionId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "profile": {
                    "type": "string"
                  },
                  "inputs": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "artifacts": {
                    "$ref": "#/components/schemas/ArtifactBindings"
                  }
                },
                "required": [
                  "definitionId",
                  "profile",
                  "inputs",
                  "artifacts"
                ]
              }
            }
          }
        }
      }
    },
    "/runs/{runId}": {
      "get": {
        "operationId": "get_runs_runId",
        "summary": "Read run with immutable snapshot and instance addresses",
        "responses": {
          "200": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "run": {
                      "$ref": "#/components/schemas/Run"
                    }
                  },
                  "required": [
                    "run"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            }
          }
        ]
      }
    },
    "/runs/{runId}/cancel": {
      "post": {
        "operationId": "post_runs_runId_cancel",
        "summary": "Request cancel without inventing client-side transitions",
        "responses": {
          "202": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedRun"
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "required": []
              }
            }
          }
        }
      }
    },
    "/runs/{runId}/resume": {
      "post": {
        "operationId": "post_runs_runId_resume",
        "summary": "Reject unsupported checkpoint resume; use an advertised resolution action",
        "responses": {
          "409": {
            "description": "UNSAFE_ACTION: no resumable checkpoint is advertised by this engine.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "required": []
              }
            }
          }
        }
      }
    },
    "/runs/{runId}/instances/{instanceId}/resolve": {
      "post": {
        "operationId": "post_runs_runId_instances_instanceId_resolve",
        "summary": "Resolve an unknown external outcome using evidence",
        "responses": {
          "202": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedRun"
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            }
          },
          {
            "name": "instanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Resolution"
              }
            }
          }
        }
      }
    },
    "/requests": {
      "get": {
        "operationId": "get_requests",
        "summary": "List saved human requests",
        "responses": {
          "200": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanRequestPage"
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/requests/{requestId}/response": {
      "post": {
        "operationId": "post_requests_requestId_response",
        "summary": "Accept first valid response to exactly this saved request",
        "responses": {
          "200": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedResponse"
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "outputs": {
                    "type": "object",
                    "additionalProperties": true
                  }
                },
                "required": [
                  "outputs"
                ]
              }
            }
          }
        }
      }
    },
    "/artifacts": {
      "get": {
        "operationId": "get_artifacts",
        "summary": "List immutable registered artifact descriptors",
        "responses": {
          "200": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactPage"
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "post_artifacts",
        "summary": "Register uploaded input bytes",
        "responses": {
          "201": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "artifact": {
                      "$ref": "#/components/schemas/Artifact"
                    }
                  },
                  "required": [
                    "artifact"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "mediaType": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string",
                    "contentEncoding": "base64"
                  }
                },
                "required": [
                  "name",
                  "mediaType",
                  "content"
                ]
              }
            }
          }
        }
      }
    },
    "/artifacts/{artifactId}": {
      "get": {
        "operationId": "get_artifacts_artifactId",
        "summary": "Read artifact identity, hash and origin",
        "responses": {
          "200": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "artifact": {
                      "$ref": "#/components/schemas/Artifact"
                    }
                  },
                  "required": [
                    "artifact"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "artifactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "diagnostics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            }
          }
        },
        "required": [
          "code",
          "message",
          "diagnostics"
        ]
      },
...
      "Info": {
        "type": "object",
        "properties": {
          "protocol": {
            "const": "knotra.desktop/1"
          },
          "engineId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          },
          "version": {
            "type": "string"
          },
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "principalId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          }
        },
        "required": [
          "protocol",
          "engineId",
          "version",
          "capabilities",
          "principalId"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Required for remote HTTPS deployments; optional loopback according to engine policy."
      }
    }
  }
}